Fix compilation on Mac.

+++ dbms/src/Storages/StorageDictionary.cpp
@@ -46,7 +46,7 @@ BlockInputStreams StorageDictionary::read(
     const SelectQueryInfo & /*query_info*/,
     const Context & context,
     QueryProcessingStage::Enum /*processed_stage*/,
-    const UInt64 max_block_size,
+    const size_t max_block_size,
     const unsigned /*threads*/)
 {
     auto dictionary = context.getExternalDictionaries().getDictionary(dictionary_name);
