+++ dbms/src/Storages/System/StorageSystemColumns.cpp
@@ -261,7 +261,7 @@ BlockInputStreams StorageSystemColumns::read(
     const SelectQueryInfo & query_info,
     const Context & context,
     QueryProcessingStage::Enum /*processed_stage*/,
-    const UInt64 max_block_size,
+    const size_t max_block_size,
     const unsigned /*num_streams*/)
 {
     check(column_names);
